Remove inheritance and use only interfaces. Add doc comments to interface and use inheritdoc.
